🚚 reorganize game code
