+++ apps/tempest.games/src/app/saloon/[roomId]/page.module.scss
@@ -2,7 +2,6 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-
   > header {
     height: 75px;
     display: inline-flex;
