+++ apps/tempest.games/src/game/game-pieces/Deck.tsx
@@ -6,38 +6,30 @@ import {
 import { AnimatePresence, motion } from "framer-motion"
 import { setCssVars } from "hamr/react-css-vars"
 import { Id } from "hamr/react-id"
-import * as React from "react"
 
-import { dealCardsTX, groupsOfCards } from "~/apps/node/lodge/src/store/game"
+import { groupsOfCards, shuffleDeckTX } from "~/apps/node/lodge/src/store/game"
 
 import { memoize } from "src/components/memoize"
 import { useRadial } from "src/services/peripherals/radial"
-import { myHandsIndex } from "src/services/store/my-hands-index"
-import { publicDeckIndex } from "src/services/store/public-deck-index"
-import { CardBack, CardFace, CardSlot } from "./Card"
-
 import { useDOMRect } from "src/services/use-dimensions"
-import { Count } from "./Count"
-import scss from "./Hand.module.scss"
+import { CardBack } from "./Card"
 
-export const Hand = memoize<{ id: string; detailed?: boolean }>(
-	`Hand`,
+import { Count } from "../labels/Count"
+import scss from "./Deck.module.scss"
+
+export const Deck = memoize<{ id: string; detailed?: string }>(
+	`Deck`,
 	({ id, detailed }) => {
-		const isMyHand = useO(myHandsIndex).includes(id)
 		const cardIds = useO(groupsOfCards.findState.cardKeysOfGroup(id))
-		const publicDeckIds = useO(publicDeckIndex)
 
 		usePullMutableFamilyMember(groupsOfCards.core.findRelatedKeysState(id))
 
-		const dealCards = useServerAction(dealCardsTX)
+		const shuffle = useServerAction(shuffleDeckTX)
 
 		const handlers = useRadial([
 			{
-				label: `Deal`,
-				do: () => {
-					const deckId = publicDeckIds[0]
-					dealCards({ deckId, handId: id, count: 1 })
-				},
+				label: `Shuffle`,
+				do: () => shuffle({ deckId: id }),
 			},
 		])
 
@@ -49,8 +41,8 @@ export const Hand = memoize<{ id: string; detailed?: boolean }>(
 				<span
 					className={scss.class}
 					style={setCssVars({
-						"--child-len": `${height * (5 / 7)}px`,
-						"--child-count": cardIds.length,
+						"--child-len": `${height}px`,
+						"--child-count": `${cardIds.length}`,
 					})}
 					{...handlers}
 				>
@@ -67,15 +59,11 @@ export const Hand = memoize<{ id: string; detailed?: boolean }>(
 						animate={{ opacity: 1 }}
 						exit={{ opacity: 0 }}
 					>
-						{cardIds.length === 0 ? (
-							<CardSlot />
-						) : isMyHand ? (
-							cardIds.map((cardId) => <CardFace key={cardId} id={cardId} />)
-						) : (
-							cardIds.map((cardId) => <CardBack key={cardId} id={cardId} />)
-						)}
-						{detailed ? null : <Count amount={cardIds.length} />}
+						{cardIds.map((cardId) => (
+							<CardBack key={cardId} id={cardId} />
+						))}
 					</motion.article>
+					{detailed ? null : <Count amount={cardIds.length} />}
 				</span>
 			</AnimatePresence>
 		)
