+++ apps/tempest.games/src/game/my-domain/Me.module.scss
@@ -4,6 +4,7 @@ span.class {
   height: 36px;
   background-color: var(--table-close);
   display: flex;
+  margin-bottom: -15px;
   h3 {
     margin: auto;
     font-size: 16px;
