+++ apps/tempest.games/src/game/my-domain/MyDomain.module.scss
@@ -7,6 +7,7 @@ div.class {
   display: flex;
   flex-flow: column;
   padding: 10px;
+  padding-bottom: 30px;
   align-items: center;
   > [data-css="me"] {
     position: relative;
