+++ apps/tempest.games/src/game/public/Public.tsx
@@ -3,7 +3,7 @@ import * as React from "react"
 import { h3 } from "src/components/<hX>"
 
 import { publicDeckIndex } from "src/services/store/public-deck-index"
-import { Deck } from "./Deck"
+import { Deck } from "../game-pieces/Deck"
 
 import scss from "./Public.module.scss"
